refactor: message resource
